More adjustments to PCA9685 servo logic.

+++ I2CRawDevice.h
@@ -57,10 +57,11 @@ public:
     PackedByteArray read_bytes_from_device( const int length );
     void write_bytes_to_device( PackedByteArray bytes );
     
-    int read_byte_from_device_register( int i2c_device_register );
-    void    write_byte_to_device_register( int i2c_device_register, int value );
-    void    write_byte_array_to_device_register( int i2c_device_register, PackedByteArray bytes );
-    void    _write_byte_array_to_device_register( uint8_t i2c_device_register, const uint8_t* bytes );
+    uint8_t read_byte_from_device_register( uint8_t i2c_device_register );
+    void    write_byte_to_device_register( uint8_t i2c_device_register, uint8_t value );
+    
+    void  write_byte_array_to_device_register( int i2c_device_register, PackedByteArray bytes );
+    void _write_byte_array_to_device_register( uint8_t i2c_device_register, const uint8_t* bytes );
 };
 
 }
